Move the prompt_for_ssh function to startup_utils
